∙ 修复视频无法匹配的问题 ∙ 修复进度条显示错误的问题 ∙ 修复文件菜单无法打开的问题 ∙ 修复自动更新出错的问题 ———————— 时隔半年 终于填坑😂😂 下版本底层会有非常大的调整 这个版本先将就用吧

+++ DanDanPlayForMac/NetManager/UpdateNetManager.m
@@ -17,7 +17,7 @@
         GDataXMLElement *rootElement = document.rootElement;
         
         VersionModel *model = [[VersionModel alloc] init];
-        model.version = [[rootElement elementsForName:@"version"].firstObject stringValue].floatValue;
+        model.version = [[rootElement elementsForName:@"version"].firstObject stringValue];
         model.details = [[rootElement elementsForName:@"details"].firstObject stringValue];
         model.md5 = [[rootElement elementsForName:@"hash"].firstObject stringValue];
         model.patchName = [[rootElement elementsForName:@"patch"].firstObject stringValue];
